feat(adicionar novo calculo de distancia)

+++ src/main/java/br/adriano/localizacaoteste/DemoApplication.java
@@ -23,16 +23,17 @@ public class DemoApplication implements CommandLineRunner{
 	
 	@Override
 	public void run(String... args) throws Exception {
-		Double lat1 = (double) -2.515776;
-		Double long1 = (double) -4.4245216;
+
+		double lat1 = -2.515776;
+		double long1 = -4.4245216;
 		Loja loja1 = new Loja(null, "Cohama", lat1, long1);
 		
-		Double lat2 = (double) -2.517882;
-		Double longi2 = (double) -4.4257721;
+		double lat2 = -2.517882;
+		double longi2 = -4.4257721;
 		Loja loja2 = new Loja(null, "Mix curva do 90", lat2, longi2);
 		
-		Double lat3 = (double)-2.527261;
-		Double longi3 = (double)-44.255085;
+		double lat3 = -2.527261;
+		double longi3 = -44.255085;
 		Loja loja3 = new Loja(null, "Shopping da ilha", lat3, longi3);
 		
 		lojaRepository.saveAll(Arrays.asList(loja1, loja2, loja3));
